Refactor game structure: update entry point to src/game.ts, implement player and enemy logic, and enhance game loop with action handling
