+++ package.json
@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "main": "src/index.ts",
   "scripts": {
-    "start": "ts-node src/index.ts",
+    "start": "ts-node src/game.ts",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
