+++ src/index.ts
@@ -1,8 +1,9 @@
 import readline from "readline";
-import { Game } from "./game";
+import { attack, defense, Game, heal } from "./game";
 
 // Initialize the game
-const game = new Game();
+const player = new Game ("DOG", 50, 10, 5);    
+const enemy = new Game ("CAT", 50, 8, 6); // Add an opponent
 
 // Setup the readline interface
 const rl = readline.createInterface({
@@ -10,21 +11,57 @@ const rl = readline.createInterface({
     output: process.stdout
 });
 
+// Random choice: 1 (Attack), 2 (Defend), 3 (Heal)
+function enemyTurn(enemy: Game, player: Game) {
+    const action = Math.floor(Math.random() * 3) + 1; 
+
+    switch (action) {
+        case 1: // Attack
+            attack(enemy, player);
+            break;
+
+        case 2: // Defend
+            defense(enemy, player);
+            break;
+
+        case 3: // Heal
+            heal(enemy);
+            break;
+    }
+}
+
 // Start the game loop
-function startGame(isFinished = false) {
-    if (isFinished) {
-        console.log("Game Over");
+function gameLoop() {
+    if (player.health <= 0) {
+        console.log("You have been defeated by a Cat!");
+        rl.close();
+        return;
+    }
+    if (enemy.health <= 0) {
+        console.log("Congratulations! You defeated the Cat!");
         rl.close();
         return;
     }
 
-    game.showActions();
+    player.showActions(player.name, player.health, player.attack, player.defense);
+    rl.question("Choose your action: ", (action: string) => {
+        if (!player.handleAction(action, enemy)) {
+            rl.close();
+            return;
+        }
 
-    rl.question("Enter your choice: ", (input: string) => {
-        isFinished = game.handleAction(input);
-        startGame();  // Loop back to show actions again
+        // Enemy turn logic
+        if (enemy.health > 0) {
+            console.log(`\nThe ${enemy.name}'s turn...`);
+            enemyTurn(enemy, player);
+        }
+        
+
+        // Show current stats
+        console.log(`\nPlayer Health: ${player.health}, Enemy Health: ${enemy.health}`);
+        gameLoop();
     });
 }
 
-console.log("Welcome to the Battle Game!");
-startGame();
+console.log("Battle begins!");
+gameLoop();
